requirements.txt
beautifulsoup4
PyPDF2
Pillow
pytesseract
pdf2image

graphviz
progress
ujson
